requirements-test.txt
requests==2.31.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytest==7.4.2
pytest-django==4.5.2
pytest-cov==4.1.0
locust==2.16.1
